refactor(devtools): mv sortFns to utils file
