+++ packages/react-query-devtools/src/devtools.tsx
@@ -12,7 +12,7 @@ import {
 } from '@tanstack/react-query'
 import { rankItem, compareItems } from '@tanstack/match-sorter-utils'
 import useLocalStorage from './useLocalStorage'
-import { useIsMounted } from './utils'
+import { sortFns, useIsMounted } from './utils'
 
 import {
   Panel,
@@ -373,27 +373,6 @@ export function ReactQueryDevtools({
       ) : null}
     </Container>
   )
-}
-
-const getStatusRank = (q: Query) =>
-  q.state.fetchStatus !== 'idle'
-    ? 0
-    : !q.getObserversCount()
-    ? 3
-    : q.isStale()
-    ? 2
-    : 1
-
-export const sortFns: Record<string, (a: Query, b: Query) => number> = {
-  'Status > Last Updated': (a, b) =>
-    getStatusRank(a) === getStatusRank(b)
-      ? (sortFns['Last Updated']?.(a, b) as number)
-      : getStatusRank(a) > getStatusRank(b)
-      ? 1
-      : -1,
-  'Query Hash': (a, b) => (a.queryHash > b.queryHash ? 1 : -1),
-  'Last Updated': (a, b) =>
-    a.state.dataUpdatedAt < b.state.dataUpdatedAt ? 1 : -1,
 }
 
 const useSubscribeToQueryCache = <T,>(
